first tests show that data is being uploaded properly to the DB. Needs further testing and might change the way I save data to use joined columns instead of referencing id manually

+++ src/main/java/ons/group8/services/ChecklistCreationEvent.java
@@ -12,5 +12,5 @@ public class ChecklistCreationEvent {
     List<Topic> topics;
     List<Long> assignedTo;
     String deadline;
-    Integer authorId;
+    Long authorId;
 }
